+++ src/tests/Recyclarr.Cli.Tests/Processors/ConfigTemplateListerTest.cs
@@ -1,4 +1,5 @@
 using System.IO.Abstractions;
+using Recyclarr.Cli.Console.Commands;
 using Recyclarr.Cli.Processors.Config;
 using Recyclarr.TrashLib.Config;
 using Recyclarr.TrashLib.Config.Services;
@@ -26,7 +27,10 @@ public class ConfigTemplateListerTest : TrashLibIntegrationFixture
             new TemplatePath {Id = "s2", TemplateFile = stubFile, Service = SupportedServices.Sonarr, Hidden = true}
         });
 
-        sut.Process();
+        var settings = Substitute.For<IConfigListTemplatesSettings>();
+        settings.Includes.Returns(false);
+
+        sut.Process(settings);
 
         console.Output.Should().NotContain("s2");
     }